requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   perplexityai
autoflake==2.2.1
    # via -r requirements.in
beautifulsoup4==4.12.2
    # via -r requirements.in
black==23.11.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
build==1.3.0
    # via pip-tools
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
cfgv==3.4.0
    # via pre-commit
click==8.1.8
    # via
    #   black
    #   flask
    #   pip-tools
distlib==0.4.0
    # via virtualenv
distro==1.9.0
    # via perplexityai
filelock==3.19.1
    # via virtualenv
flake8==6.1.0
    # via -r requirements.in
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==4.0.0
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
gunicorn==21.2.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.25.2
    # via
    #   -r requirements.in
    #   perplexityai
hyperframe==6.1.0
    # via h2
identify==2.6.14
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.1.0
    # via pytest
isort==5.12.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
mccabe==0.7.0
    # via flake8
mypy-extensions==1.1.0
    # via black
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   black
    #   build
    #   gunicorn
    #   pytest
pathspec==0.12.1
    # via black
perplexityai==0.20.0
    # via -r requirements.in
pip-tools==7.5.0
    # via -r requirements.in
platformdirs==4.4.0
    # via
    #   black
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==3.6.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pycodestyle==2.11.1
    # via flake8
pydantic==2.10.0
    # via
    #   -r requirements.in
    #   perplexityai
pydantic-core==2.27.0
    # via pydantic
pyflakes==3.1.0
    # via
    #   autoflake
    #   flake8
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==7.4.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0.3
    # via pre-commit
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   perplexityai
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.23
    # via
    #   -r requirements.in
    #   alembic
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
virtualenv==20.34.0
    # via pre-commit
werkzeug==3.1.3
    # via flask
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
